add Dart compilation support
