+++ web/new_playground.html
@@ -75,7 +75,11 @@
         </div>
         <div id="editor-host"></div>
     </div>
-    <div id="output-panel"></div>
+    <div id="output-panel">
+        <iframe id="frame" sandbox="allow-scripts" flex src="scripts/frame_dark.html">
+        </iframe>
+        <div id="output-host"></div>
+    </div>
 </section>
 
 <footer>
